Fixed bug where 6x6 game gets stuck
There was a logic error in game controller for cases where the AI player has no available moves but the human player does.  In that case, the human player is supposed to keep on moving until a move becomes available to the AI player or the game ends.  Fixed a couple lines of code to get that working as it should.